+++ qiskit/extensions/standard/rx.py
@@ -34,7 +34,7 @@ class RXGate(Gate):
 
     def reapply(self, circ):
         """Reapply this gate to corresponding qubits in circ."""
-        self._modifiers(circ.rx(self.param[0], self.arg[0]))
+        self._modifiers(circ.rx(self.param[0], self.qargs[0]))
 
 
 def rx(self, theta, q):