+++ qiskit/extensions/standard/ry.py
@@ -34,7 +34,7 @@ class RYGate(Gate):
 
     def reapply(self, circ):
         """Reapply this gate to corresponding qubits in circ."""
-        self._modifiers(circ.ry(self.param[0], self.arg[0]))
+        self._modifiers(circ.ry(self.param[0], self.qargs[0]))
 
 
 def ry(self, theta, q):